Move to org-mode 8.0
Since this is a breaking change I've taken the chance to clean up my
configuration a bit while I'm at it.

The other big change is the new exporter. Over riding a just the
template function is now straight forward. This means that I can define
a function that builds my site just as I like it. Fantastic.